use header compare_less_visitor
