+++ include/numsim_cas/compare_equal_visitor.h
@@ -1,0 +1,4 @@
+#ifndef COMPARE_EQUAL_VISITOR_H
+#define COMPARE_EQUAL_VISITOR_H
+
+#endif // COMPARE_EQUAL_VISITOR_H
